+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## Not yet released
 
+  * update: New module for upgrade from Magento 2.1 to 2.2 added (use only for upgrade)
   * update: Magento 2.2 compatibility (compatibility preserved with previous versions)
   * update: Family demo file updated
   * update: fix magic number